리팩토링: sql 용어 변경
