+++ src/main/java/com/example/finalproject/domain/elasticsearchpopular/service/PopularSearchIndexer.java
@@ -2,7 +2,7 @@ package com.example.finalproject.domain.elasticsearchpopular.service;
 
 import co.elastic.clients.elasticsearch.ElasticsearchClient;
 import com.example.finalproject.domain.elasticsearchpopular.dto.PopularSearchDoc;
-import com.example.finalproject.domain.elasticsearchpopular.entity.PopularSearch;
+import com.example.finalproject.domain.elasticsearchpopular.entity.PopularSearches;
 import com.example.finalproject.domain.elasticsearchpopular.repository.PopularSearchRepository;
 import io.jsonwebtoken.io.IOException;
 import lombok.RequiredArgsConstructor;
@@ -26,19 +26,19 @@ public class PopularSearchIndexer {
         // 1. 모든 지역 조회 (DB에 있는 distinct 지역)
         List<String> regions = repository.findAll()
                 .stream()
-                .map(PopularSearch::getRegion)
+                .map(PopularSearches::getRegion)
                 .distinct()
                 .toList();
 
         // 2. 지역별 Top10 가져와서 Elasticsearch에 색인
         for (String region : regions) {
-            List<PopularSearch> topSearches = repository.findTop10ByRegionOrderByCountDesc(region);
+            List<PopularSearches> topSearches = repository.findTop10ByRegionOrderByCountDesc(region);
 
-            for (PopularSearch search : topSearches) {
+            for (PopularSearches search : topSearches) {
                 PopularSearchDoc doc = PopularSearchDoc.builder()
                         .keyword(search.getKeyword())
                         .region(search.getRegion())
-                        .count(search.getCount())
+                        .search_count(search.getSearch_count())
                         .createdAt(search.getCreatedAt())
                         .build();
 
@@ -49,7 +49,7 @@ public class PopularSearchIndexer {
                         .document(Map.of(
                                 "region", search.getRegion(),
                                 "keyword", search.getKeyword(),
-                                "count", search.getCount(),
+                                "search_count", search.getSearch_count(),
                                 "created_at", search.getCreatedAt(),
                                 "type", "db" //DB -> ES 인덱싱, 문서 생성
                         ))
